test-requirements.txt
selenium==2.42.1
requests==2.3.0
pytest
pytest-django==2.8.0
mock==1.0.1
